Adds trailing slash to site base url
This fixes issues.

When running the hugo dev server it adds it to local host. Thisngs that worked locally were broken when built for the server if paths didn’t start with a leading /.

This brings them in line and prevents the need to start all local path urls with a slash.

Signed-off-by: Victor Hudson <vichudson1@gmail.com>